+++ packages/components-angular/src/app/examples/index.ts
@@ -1,6 +1,7 @@
 import { AccordionExampleComponent } from './accordion-example.component';
 import { BannerExampleComponent } from './banner-example.component';
 import { ButtonExampleFormComponent } from './button-example-form.component';
+import { CanvasExampleComponent } from './canvas-example.component';
 import { CarouselExampleDynamicSlidesComponent } from './carousel-example-dynamic-slides.component';
 import { CarouselExampleEventsComponent } from './carousel-example-events.component';
 import { CarouselExampleJumpToSlideComponent } from './carousel-example-jump-to-slide.component';
@@ -39,6 +40,7 @@ export const examples = [
   AccordionExampleComponent,
   BannerExampleComponent,
   ButtonExampleFormComponent,
+  CanvasExampleComponent,
   CarouselExampleDynamicSlidesComponent,
   CarouselExampleEventsComponent,
   CarouselExampleJumpToSlideComponent,
@@ -77,6 +79,7 @@ export const examples = [
 export * from './accordion-example.component';
 export * from './banner-example.component';
 export * from './button-example-form.component';
+export * from './canvas-example.component';
 export * from './carousel-example-dynamic-slides.component';
 export * from './carousel-example-events.component';
 export * from './carousel-example-jump-to-slide.component';