+++ packages/components-react/projects/react-wrapper/src/BaseProps.ts
@@ -1,0 +1,23 @@
+import type { DOMAttributes, HTMLAttributes } from 'react';
+
+// handpicked props that all wrapper components have in common
+export type BaseProps = DOMAttributes<{}> &
+  Pick<
+    HTMLAttributes<{}>,
+    // React-specific Attributes
+    | 'suppressHydrationWarning'
+    // Standard HTML Attributes
+    | 'autoFocus'
+    | 'className'
+    | 'dir'
+    | 'hidden'
+    | 'id'
+    | 'lang'
+    | 'slot'
+    | 'style'
+    | 'tabIndex'
+    | 'title'
+    | 'translate'
+    // WAI-ARIA
+    | 'role'
+  >;