+++ packages/components/src/components.d.ts
@@ -12,6 +12,7 @@ import { ButtonIcon } from "./components/button/button-utils";
 import { ButtonGroupDirection } from "./components/button-group/button-group-utils";
 import { ButtonPureAlignLabel, ButtonPureAriaAttribute, ButtonPureIcon, ButtonPureSize, ButtonPureType, ButtonPureWeight } from "./components/button-pure/button-pure-utils";
 import { ButtonTileAlign, ButtonTileAriaAttribute, ButtonTileAspectRatio, ButtonTileBackground, ButtonTileIcon, ButtonTileSize, ButtonTileType, ButtonTileWeight } from "./components/button-tile/button-tile-utils";
+import { CanvasSidebarEndWidth, CanvasSidebarStartWidth } from "./components/canvas/canvas-utils";
 import { CarouselAlignHeader, CarouselAriaAttribute, CarouselHeadingSize, CarouselInternationalization, CarouselUpdateEventDetail, CarouselWidth } from "./components/carousel/carousel-utils";
 import { CheckboxWrapperState } from "./components/checkbox-wrapper/checkbox-wrapper-utils";
 import { ContentWrapperBackgroundColor, ContentWrapperWidth } from "./components/content-wrapper/content-wrapper-utils";
@@ -47,7 +48,7 @@ import { PopoverAriaAttribute, PopoverDirection } from "./components/popover/pop
 import { RadioButtonWrapperState } from "./components/radio-button-wrapper/radio-button-wrapper-utils";
 import { ScrollerAlignScrollIndicator, ScrollerAriaAttribute, ScrollerGradientColor, ScrollerGradientColorScheme, ScrollerScrollIndicatorPosition, ScrollerScrollToPosition } from "./components/scroller/scroller-utils";
 import { SegmentedControlBackgroundColor, SegmentedControlColumns, SegmentedControlUpdateEventDetail } from "./components/segmented-control/segmented-control/segmented-control-utils";
-import { SegmentedControlItemIcon } from "./components/segmented-control/segmented-control-item/segmented-control-item-utils";
+import { SegmentedControlItemAriaAttribute, SegmentedControlItemIcon } from "./components/segmented-control/segmented-control-item/segmented-control-item-utils";
 import { SelectDropdownDirection, SelectState, SelectUpdateEventDetail } from "./components/select/select/select-utils";
 import { SelectWrapperDropdownDirection, SelectWrapperState } from "./components/select-wrapper/select-wrapper/select-wrapper-utils";
 import { SpinnerAriaAttribute, SpinnerSize } from "./components/spinner/spinner-utils";
@@ -73,6 +74,7 @@ export { ButtonIcon } from "./components/button/button-utils";
 export { ButtonGroupDirection } from "./components/button-group/button-group-utils";
 export { ButtonPureAlignLabel, ButtonPureAriaAttribute, ButtonPureIcon, ButtonPureSize, ButtonPureType, ButtonPureWeight } from "./components/button-pure/button-pure-utils";
 export { ButtonTileAlign, ButtonTileAriaAttribute, ButtonTileAspectRatio, ButtonTileBackground, ButtonTileIcon, ButtonTileSize, ButtonTileType, ButtonTileWeight } from "./components/button-tile/button-tile-utils";
+export { CanvasSidebarEndWidth, CanvasSidebarStartWidth } from "./components/canvas/canvas-utils";
 export { CarouselAlignHeader, CarouselAriaAttribute, CarouselHeadingSize, CarouselInternationalization, CarouselUpdateEventDetail, CarouselWidth } from "./components/carousel/carousel-utils";
 export { CheckboxWrapperState } from "./components/checkbox-wrapper/checkbox-wrapper-utils";
 export { ContentWrapperBackgroundColor, ContentWrapperWidth } from "./components/content-wrapper/content-wrapper-utils";
@@ -108,7 +110,7 @@ export { PopoverAriaAttribute, PopoverDirection } from "./components/popover/pop
 export { RadioButtonWrapperState } from "./components/radio-button-wrapper/radio-button-wrapper-utils";
 export { ScrollerAlignScrollIndicator, ScrollerAriaAttribute, ScrollerGradientColor, ScrollerGradientColorScheme, ScrollerScrollIndicatorPosition, ScrollerScrollToPosition } from "./components/scroller/scroller-utils";
 export { SegmentedControlBackgroundColor, SegmentedControlColumns, SegmentedControlUpdateEventDetail } from "./components/segmented-control/segmented-control/segmented-control-utils";
-export { SegmentedControlItemIcon } from "./components/segmented-control/segmented-control-item/segmented-control-item-utils";
+export { SegmentedControlItemAriaAttribute, SegmentedControlItemIcon } from "./components/segmented-control/segmented-control-item/segmented-control-item-utils";
 export { SelectDropdownDirection, SelectState, SelectUpdateEventDetail } from "./components/select/select/select-utils";
 export { SelectWrapperDropdownDirection, SelectWrapperState } from "./components/select-wrapper/select-wrapper/select-wrapper-utils";
 export { SpinnerAriaAttribute, SpinnerSize } from "./components/spinner/spinner-utils";
@@ -381,6 +383,27 @@ export namespace Components {
           * Font weight of the description.
          */
         "weight"?: BreakpointCustomizable<ButtonTileWeight>;
+    }
+    /**
+     * @experimental 
+     */
+    interface PCanvas {
+        /**
+          * Open Sidebar on the end side
+         */
+        "sidebarEndOpen"?: boolean;
+        /**
+          * Defines the width of the sidebar on the end side
+         */
+        "sidebarEndWidth"?: CanvasSidebarEndWidth;
+        /**
+          * Open Sidebar on the start side
+         */
+        "sidebarStartOpen"?: boolean;
+        /**
+          * Defines the width of the sidebar on the start side
+         */
+        "sidebarStartWidth"?: CanvasSidebarStartWidth;
     }
     interface PCarousel {
         /**
@@ -1504,6 +1527,10 @@ export namespace Components {
     }
     interface PSegmentedControlItem {
         /**
+          * Add ARIA attributes.
+         */
+        "aria"?: SelectedAriaAttributes<SegmentedControlItemAriaAttribute>;
+        /**
           * Disables the button. No events will be triggered while disabled state is active.
          */
         "disabled"?: boolean;
@@ -2022,6 +2049,10 @@ export interface PBannerCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPBannerElement;
 }
+export interface PCanvasCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLPCanvasElement;
+}
 export interface PCarouselCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLPCarouselElement;
@@ -2153,6 +2184,27 @@ declare global {
     var HTMLPButtonTileElement: {
         prototype: HTMLPButtonTileElement;
         new (): HTMLPButtonTileElement;
+    };
+    interface HTMLPCanvasElementEventMap {
+        "dismissSidebarStart": void;
+        "dismissSidebarEnd": void;
+    }
+    /**
+     * @experimental 
+     */
+    interface HTMLPCanvasElement extends Components.PCanvas, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLPCanvasElementEventMap>(type: K, listener: (this: HTMLPCanvasElement, ev: PCanvasCustomEvent<HTMLPCanvasElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLPCanvasElementEventMap>(type: K, listener: (this: HTMLPCanvasElement, ev: PCanvasCustomEvent<HTMLPCanvasElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLPCanvasElement: {
+        prototype: HTMLPCanvasElement;
+        new (): HTMLPCanvasElement;
     };
     interface HTMLPCarouselElementEventMap {
         "carouselChange": CarouselUpdateEventDetail;
@@ -2796,6 +2848,7 @@ declare global {
         "p-button-group": HTMLPButtonGroupElement;
         "p-button-pure": HTMLPButtonPureElement;
         "p-button-tile": HTMLPButtonTileElement;
+        "p-canvas": HTMLPCanvasElement;
         "p-carousel": HTMLPCarouselElement;
         "p-checkbox-wrapper": HTMLPCheckboxWrapperElement;
         "p-content-wrapper": HTMLPContentWrapperElement;
@@ -3129,6 +3182,35 @@ declare namespace LocalJSX {
           * Font weight of the description.
          */
         "weight"?: BreakpointCustomizable<ButtonTileWeight>;
+    }
+    /**
+     * @experimental 
+     */
+    interface PCanvas {
+        /**
+          * Emitted when the component requests to close the sidebar on the end side.
+         */
+        "onDismissSidebarEnd"?: (event: PCanvasCustomEvent<void>) => void;
+        /**
+          * Emitted when the component requests to close the sidebar on the start side.
+         */
+        "onDismissSidebarStart"?: (event: PCanvasCustomEvent<void>) => void;
+        /**
+          * Open Sidebar on the end side
+         */
+        "sidebarEndOpen"?: boolean;
+        /**
+          * Defines the width of the sidebar on the end side
+         */
+        "sidebarEndWidth"?: CanvasSidebarEndWidth;
+        /**
+          * Open Sidebar on the start side
+         */
+        "sidebarStartOpen"?: boolean;
+        /**
+          * Defines the width of the sidebar on the start side
+         */
+        "sidebarStartWidth"?: CanvasSidebarStartWidth;
     }
     interface PCarousel {
         /**
@@ -4316,6 +4398,10 @@ declare namespace LocalJSX {
     }
     interface PSegmentedControlItem {
         /**
+          * Add ARIA attributes.
+         */
+        "aria"?: SelectedAriaAttributes<SegmentedControlItemAriaAttribute>;
+        /**
           * Disables the button. No events will be triggered while disabled state is active.
          */
         "disabled"?: boolean;
@@ -4883,6 +4969,7 @@ declare namespace LocalJSX {
         "p-button-group": PButtonGroup;
         "p-button-pure": PButtonPure;
         "p-button-tile": PButtonTile;
+        "p-canvas": PCanvas;
         "p-carousel": PCarousel;
         "p-checkbox-wrapper": PCheckboxWrapper;
         "p-content-wrapper": PContentWrapper;
@@ -4960,6 +5047,10 @@ declare module "@stencil/core" {
             "p-button-group": LocalJSX.PButtonGroup & JSXBase.HTMLAttributes<HTMLPButtonGroupElement>;
             "p-button-pure": LocalJSX.PButtonPure & JSXBase.HTMLAttributes<HTMLPButtonPureElement>;
             "p-button-tile": LocalJSX.PButtonTile & JSXBase.HTMLAttributes<HTMLPButtonTileElement>;
+            /**
+             * @experimental 
+             */
+            "p-canvas": LocalJSX.PCanvas & JSXBase.HTMLAttributes<HTMLPCanvasElement>;
             "p-carousel": LocalJSX.PCarousel & JSXBase.HTMLAttributes<HTMLPCarouselElement>;
             "p-checkbox-wrapper": LocalJSX.PCheckboxWrapper & JSXBase.HTMLAttributes<HTMLPCheckboxWrapperElement>;
             /**