+++ packages/components/src/components/flyout-navigation/flyout-navigation-item/flyout-navigation-item.tsx
@@ -69,6 +69,9 @@ export class FlyoutNavigationItem {
         </PrefixedTagNames.pButtonPure>
         <div
           class="scroller"
+          // "inert" will be known from React 19 onwards, see https://github.com/facebook/react/pull/24730
+          // eslint-disable-next-line
+          /* @ts-ignore */
           inert={this.open ? null : true} // prevents focusable elements during fade-out transition
         >
           <div class="header">